+++ src/aks-preview/HISTORY.md
@@ -2,10 +2,18 @@
 
 Release History
 ===============
+
+0.5.6
++++++
+* Fix issue that assigning identity in another subscription will fail
+
+0.5.5
++++++
+* Add support for Azure KeyVault Secrets Provider as an AKS addon
+
 0.5.4
 +++++
 * Add operations of maintenance configuration
-=======
 
 0.5.3
 +++++